+++ config.env
@@ -1,2 +1,5 @@
 NODE_ENV=development
 
+export DATABASE_URL=postgres://wpsukukdaptnbm:e1ffadb271aca96da3036a9b924fc729b109f16f54da2d7653cdba4c78fce2ec@ec2-54-195-241-106.eu-west-1.compute.amazonaws.com:5432/d9th13b3ld2pt7
+
+export DB_URLTEST=postgres://tester:password@localhost:5432/testDB
