+++ package.json
@@ -4,8 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "node src/tests/*.test.js | tap-spec",
-    "dbtest": "NODE_ENV=test node src/tests/*.test.js | tap-spec",
+    "test": "NODE_ENV=test node src/tests/*.test.js | tap-spec",
     "nodemon": "nodemon src/index.js",
     "start": "node src/index.js"
   },
