+++ src/tests/queries.test.js
@@ -1,11 +1,36 @@
 const test = require('tape');
 const runDbBuild = require('../database/dbBuild.js');
 const queries = require('../controllers/queries');
+const request = require('supertest');
+const app = require('../app');
 
-test('testing tape is working', (t) => {
-  t.equals(1, 1, 'one equals one');
-  t.end();
-});
+//TESTING ROUTES:
+
+
+const endpoints = ['/',
+           '/search',
+           '/search/:category',
+           '/place/:place',
+           '/add',
+           '/add/checkExists',
+           '/add/details'
+           ];
+
+
+endpoints.forEach(endpoint => {
+  test('All routes should return status code 200 and text/html', t => {
+    request(app)
+      .get(`${endpoint}`)
+      .expect(200)
+      .expect('Content-Type', 'text/html; charset=utf-8')
+      .end(err => {
+        t.error(err);
+        t.end();
+      })
+  })
+})
+
+//TESTING QUERIES:
 
 test('testing getPlace query returns an object', (t) => {
   runDbBuild((err, res) => {
